Extract `Description` component for usage in other situations
